--- a/conn_test.go
+++ b/conn_test.go
@@ -620,7 +620,7 @@ func BenchmarkBroadcastNoCompressionPrepared(b *testing.B) {
 	for j := 0; j < b.N; j++ {
 		for i := 0; i < bench.numMessages; i++ {
 			msg := bench.messages[i%len(bench.messages)]
-			preparedMsg, _ := NewPreparedMessage(TextMessage, msg, false, 1)
+			preparedMsg := NewPreparedMessage(TextMessage, msg)
 			for _, c := range conns {
 				c.messages <- preparedMsg
 			}
@@ -638,7 +638,7 @@ func BenchmarkBroadcastWithCompressionPrepared(b *testing.B) {
 	for j := 0; j < b.N; j++ {
 		for i := 0; i < bench.numMessages; i++ {
 			msg := bench.messages[i%len(bench.messages)]
-			preparedMsg, _ := NewPreparedMessage(TextMessage, msg, true, 1)
+			preparedMsg := NewPreparedMessage(TextMessage, msg)
 			for _, c := range conns {
 				c.messages <- preparedMsg
 			}
@@ -655,7 +655,7 @@ func TestPreparedMessageBytesStreamUncompressed(t *testing.T) {
 	var b1 bytes.Buffer
 	c := newConn(fakeNetConn{Reader: nil, Writer: &b1}, true, 1024, 1024)
 	for _, msg := range messages {
-		preparedMsg, _ := NewPreparedMessage(TextMessage, msg, false, 1)
+		preparedMsg := NewPreparedMessage(TextMessage, msg)
 		c.WritePreparedMessage(preparedMsg)
 	}
 	out1 := b1.Bytes()
@@ -679,8 +679,10 @@ func TestPreparedMessageBytesStreamCompressed(t *testing.T) {
 	c := newConn(fakeNetConn{Reader: nil, Writer: &b1}, true, 1024, 1024)
 	c.enableWriteCompression = true
 	c.newCompressionWriter = compressNoContextTakeover
-	for _, msg := range messages {
-		preparedMsg, _ := NewPreparedMessage(TextMessage, msg, true, 1)
+	for i, msg := range messages {
+		preparedMsg := NewPreparedMessage(TextMessage, msg)
+		level := i%(maxCompressionLevel-minCompressionLevel+1) - 2
+		c.SetCompressionLevel(level)
 		c.WritePreparedMessage(preparedMsg)
 	}
 	out1 := b1.Bytes()
@@ -689,7 +691,9 @@ func TestPreparedMessageBytesStreamCompressed(t *testing.T) {
 	c = newConn(fakeNetConn{Reader: nil, Writer: &b2}, true, 1024, 1024)
 	c.enableWriteCompression = true
 	c.newCompressionWriter = compressNoContextTakeover
-	for _, msg := range messages {
+	for i, msg := range messages {
+		level := i%(maxCompressionLevel-minCompressionLevel+1) - 2
+		c.SetCompressionLevel(level)
 		c.WriteMessage(TextMessage, msg)
 	}
 	out2 := b2.Bytes()