This is a concept of `PreparedMessage` seen in #182 to improve memory usage with enabled compression mentioned in #203 

I don't pretend that this pull request will be merged in a form as it's now. It's just an attempt to prove that preparing message can help a lot to reduce CPU and memory usage in PUB/SUB scenarios with many active subscribers (broadcast) especially when permessage compression is enabled.

I wrote several benchmarks inside - create many (10k) connections and send the same message to them:

* without compression
* with enabled compression
* without compression preparing message before broadcasting
* with compression enabled preparing message before broadcasting

It's possible to tweak some constants in benchmarks to modify behaviour a bit (num of conns, num of messages on every iteration).

Here results from my machine:

```
fz@websocket: go test -test.run=XXX -bench=BenchmarkBroadcast
BenchmarkBroadcastNoCompression-4             	     200	   8363120 ns/op	       9 B/op	       0 allocs/op
BenchmarkBroadcastWithCompression-4           	      50	  84743048 ns/op	101941565 B/op	   31711 allocs/op
BenchmarkBroadcastNoCompressionPrepared-4     	     200	   8315049 ns/op	     321 B/op	       2 allocs/op
BenchmarkBroadcastWithCompressionPrepared-4   	     200	   8447478 ns/op	   20435 B/op	       5 allocs/op
```

The difference is huge! This makes me feel that I did something wrong:) But at least I wrote tests to compare connection streams when using `PreparedMessage` and not. Also tried on simple chat demo - it works.

@garyburd as you can see I make `PreparedMessage` using fake connection writers - it was the simplest solution I could imagine that does not require massive refactoring and a bigger chance for me to make a mistake.