@garyburd 2017-01-24
I took a quick look at this. I might be mistaken, but it looks like PreparedMessage payload fields share memory with pooled connections. 

The fake connection is clever.  I don't think the a massive refactoring is required, but I would need to spend some time on it to know for sure.

The NextWriter/Write/Close sequence can be replaced with a call to WriteMessage.

Some type assertions can be eliminated by declaring netConn as

    type netConn struct {
        bytes.Buffer
    }
    func (netConn) Read(p []byte) (int, error) { return 0, nil }