@garyburd 2017-01-24
I noticed that the client case is not handled. Window sizes will not be handled when and if that feature is added. To handle all of the possible message variants, I suggest creating the variants lazily from an uncompressed server message.

     type preparedKey struct {
         isServer bool
         compressed bool
    }

    type preparedFrame struct {
        once sync.Once
        data []byte
    }
   
    type PreparedMessage struct {
        frameType int
        data []byte // slice of frames[preparedKey{isServer: true, compressed: false}].data
        mu sync.Mutex
        frames map[frameKey]*preparedFrame
    }

    func (pm *PreparedMessage) frameData(c *Conn) []byte {
       key := compute from connection
       pm.mu.Lock()
       frame := pm.frames[key]
       if frame == nil {
           frame = &preparedFrame{}
           pm.frames[key] = frame
       } 
       pm.mu.Unlock()
       frame.once.Do(func() {
             // create frame.data using key and pm.data
       })
       return frame.data
    }

    func (c *Conn) WritePreparedMessage(pm *PreparedMessage) error {
       data := pm.frameData(c)
       if c.isWriting {
          panic("concurrent write to websocket connection")
       }
       c.isWriting = true
       err := c.write(pm.frameType, c.writeDeadline, data, nil)
       if !c.isWriting {
              panic("concurrent write to websocket connection")
       }
       c.isWriting = false
       return err
    }

The options for NewPreparedMessage should be passed as a struct to make it possible to add more options later:

    func NewPreparedMessage(messageType int, data []byte, options *PreparedMessageOptions) *PreparedMessage {
    }

    type PreparedMessageOptions struct {
       compressionLevel int // zero means use default, not no compression
    }