@FZambia 2017-01-24
Thanks for looking!

> I might be mistaken, but it looks like PreparedMessage payload fields share memory with pooled connections.

Exactly, need a copy

> To handle all of the possible message variants, I suggest creating the variants lazily from an uncompressed server message.

Looks like `preparedKey` should also include compression level as it can be set per connection right?