@garyburd 2017-01-24
> Looks like preparedKey should also include compression level as it can be set per connection right?

I specified the compression level in the prepared message options because I think that applications will select the compression level based on the size and type of the message data.  I don't think it's something that applications will want to set per connection.

The connection API has SetCompressionLevel to match SetWriteDeadline and EnableCompression. If I could design the API from scratch, I'd put the options in a struct:

     type WriteOptions {
          messageType int
          timeout time.Duration
          disableCompression bool
          compressionLevel int // 0 uses default level
    }

     func (c *Conn) NextWriter(options *WriteOptions) (io.WriteCloser, error) {
     }

    func NewPreparedMessage(data []byte, options *WriteOptions) *PreparedMessage {
    }