@garyburd 2017-01-24
As I think about it more, I don't like that prepared messages in my proposal use SetWriteDeadline and ignores EnableCompression and SetCompressionLevel.  These are details that the developer should not need to think about.  I think it's simpler to say that the prepared message is sent using the connection's current settings.  In that case, compression level should be part of the key.

I'd like to find a better name than "prepared" for the feature because the feature is really a cache.  Perhaps something like this:

    // MultiMessage caches on the wire representations of a message payload.
    // Use MultiMessage to efficiently send a message payload to multiple
    // connections. MultiMessage is especially useful when compression 
    / is used because the CPU and memory expensive compression operation
    // can be executed once for a given set of compression options. 
    type MultiMessage struct {
    }

    func NewMultiMessage(messageType int, data []byte) *MultiMessage {
    }

I am not happy with this name either, but I think it's better than PreparedMessage.