@FZambia 2017-01-24
@garyburd I'll try to implement everything you said. What's good in `PreparedMessage` name is that it's already used in websocket libraries for other programming languages - so it could be convenient for some developers. If you don't mind I'll leave it unchanged until your next pr revision and then will change on final variant.