@garyburd 2017-01-24
Sounds good.  Do we agree on the following API?

    type PreparedMessag struct { /* unexported fields */ }
     func NewPreparedMessage(messageType int, data []byte) *PreparedMessage
     func (c *Conn) WritePreparedMessage(*PreparedMessage) error

If you would like, I'll take on refactoring the code to eliminate the need for fake conn.  I can do this after you are done with everything else.