@FZambia 2017-01-24
Yes, let me then push everything we discussed and then you can take over on this.