@FZambia 2017-01-24
@garyburd just pushed all changes - also added test that requires fix in #208 to pass.