@FZambia 2017-01-24
Oh, well, this won't pass also on old go versions because I use `-2` compression level in `TestPreparedMessageBytesStreamCompressed` test. But as you will anyway refactoring/renaming things please change this as you like. I won't touch this pr - looking forward to your changes, if you want you can even close this and start from scratch.