@FZambia 2017-02-02
@garyburd did you have a chance to turn to this since then? Maybe I can help with sth - at least fix tests and shadowing?